Reorganize integration tests
